read desc, wifi_hat_tracking_servo.py is main executible
Almost finished system that tracks, analyses, and responds accordingly to captured web camera.

As of right now, a laptop is used to do external calculations since It is using the trained ai model.

It is getting input from the esp32-cam module via a web server and sending a  custom JSON file to the other esp32 via the connected wifi.

Known issues as of now:
The cam module is very laggy, though this is most likely a network issue and cannot really be solved without a proper  antenna